♻️ Merge Rollout.Comparator into Config.UserComparator
